x/perp/types/message_withdraw_cross_margin_test.go
package types

import (
	"testing"

	"gluon/testutil/sample"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgWithdrawCrossMargin_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgWithdrawCrossMargin
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgWithdrawCrossMargin{
				User: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgWithdrawCrossMargin{
				User: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
